fix: Change device_code interval from i64 to i32
The migration creates the interval column as integer (INT4) but the
entity and storage struct used i64 (INT8), causing a type mismatch
error on PostgreSQL.

+++ src/entities/device_code.rs
@@ -14,7 +14,7 @@ pub struct Model {
     pub created_at: i64,
     pub expires_at: i64,
     pub last_poll_at: Option<i64>,
-    pub interval: i64,
+    pub interval: i32,
     pub status: String, // "pending" | "approved" | "denied" | "consumed"
     pub subject: Option<String>,
     pub auth_time: Option<i64>,
